requirements.txt
streamlit
yfinance
openai
requests
python-dotenv
serpapi